Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 int main(){
     float x,y;
     scanf("%f %f",&x,&y);
-    float product= x*y;
+    int product= x*y;
     printf("Product: %f",product);
     return 0;
 }
